Improve netif guessing using macaddr and variants
It's not rare that LLDP returns a mac address as port name. If the
neighbour has been found on netbox, and only one interface has this
macaddr, we can consider that this interface is the one we were looking
for.

Also, variants are tested on the LLDP ifname and netbox interfaces,
instead of just the netbox interfaces.