Fix missed interfaces when cleaning unmatched
When cleaning unmatched interfaces from what has been fetched from the
device, and what is stored in netbox, some of them were not checked. It
was due to the offset not being refreshed after deleting an interface,
imposing a larger shift than necessarily for the next query.

Force to fetch all interfaces locally instead of rely on a generator.